SentiVision/requirements.txt
tensorflow>=2.10.0
google-cloud-firestore>=2.11.0
google-auth>=2.17.0
opencv-python>=4.7.0
numpy>=1.21.0
pytz>=2023.3
einops>=0.6.0
